[popups] Apply aria-hidden to Arrow parts
